requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
requests==2.32.3
python-dotenv==1.1.0
pydantic==2.10.3 